# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Operator Companion edge-api",
    "version": "0.5.0"
  },
  "paths": {
...
    "/system/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "System Status",
        "operationId": "system_status_system_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/hazards": {
      "get": {
        "tags": [
          "hazards"
        ],
        "summary": "List Hazards",
        "operationId": "list_hazards_hazards_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minx,miny,maxx,maxy in local metres",
              "title": "Bbox"
            },
            "description": "minx,miny,maxx,maxy in local metres"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Hazards Hazards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hazards"
        ],
        "summary": "Create Hazard",
        "operationId": "create_hazard_hazards_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HazardIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Hazard Hazards Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hazards/{pin_id}": {
      "patch": {
        "tags": [
          "hazards"
        ],
        "summary": "Patch Hazard",
        "operationId": "patch_hazard_hazards__pin_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pin Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Hazard Hazards  Pin Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "hazards"
        ],
        "summary": "Delete Hazard",
        "operationId": "delete_hazard_hazards__pin_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Hazard Hazards  Pin Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Geometry": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Point",
              "Polygon"
            ],
            "title": "Type"
          },
          "coordinates": {
            "items": {},
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "Geometry",
        "description": "GeoJSON-shaped Point or Polygon in local site metres."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HazardIn": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "WORKER_ZONE",
              "OVERHEAD_LINE",
              "SOFT_GROUND",
              "BURIED_UTILITY",
              "TRENCH",
              "DROP_OFF",
              "OTHER"
            ],
            "title": "Type"
          },
          "geometry": {
            "$ref": "#/components/schemas/Geometry"
          },
          "radius_m": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius M"
          },
          "line_clearance_m": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Clearance M"
          }
        },
        "type": "object",
        "required": [
          "type",
          "geometry"
        ],
        "title": "HazardIn"
      },
      "LoginRequest": {
        "properties": {
          "badge_id": {
            "type": "string",
            "title": "Badge Id"
          },
          "pin": {
            "type": "string",
            "title": "Pin"
          },
          "machine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Id"
          }
        },
        "type": "object",
        "required": [
          "badge_id",
          "pin"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "operator": {
            "$ref": "#/components/schemas/OperatorOut"
          },
          "shift": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "operator"
        ],
        "title": "LoginResponse"
      },
      "MachineStatus": {
        "properties": {
          "state_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Class"
          },
          "data_stale": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Stale"
          },
          "sensor_health": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Health"
          },
          "disabled_rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Disabled Rules"
          },
          "alerts_per_operating_hour": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerts Per Operating Hour"
          },
          "dropped_frames": {
            "type": "integer",
            "title": "Dropped Frames"
          },
          "restarts": {
            "type": "integer",
            "title": "Restarts"
          },
          "last_restart_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Restart S"
          }
        },
        "type": "object",
        "required": [
          "state_class",
          "data_stale",
          "sensor_health",
          "disabled_rules",
          "alerts_per_operating_hour",
          "dropped_frames",
          "restarts",
          "last_restart_s"
        ],
        "title": "MachineStatus"
      },
...
      "OperatorOut": {
        "properties": {
          "operator_id": {
            "type": "string",
            "title": "Operator Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "operator_id",
          "name",
          "role"
        ],
        "title": "OperatorOut"
      },
      "Patch": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "CONFIRM",
              "RESOLVE"
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "Patch"
      },
...
      "SystemStatus": {
        "properties": {
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "machines": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MachineStatus"
            },
            "type": "object",
            "title": "Machines"
          },
          "mqtt": {
            "additionalProperties": true,
            "type": "object",
            "title": "Mqtt"
          },
          "sync": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sync"
          },
          "models": {
            "additionalProperties": true,
            "type": "object",
            "title": "Models"
          },
          "sim": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sim"
          },
          "auth_disabled": {
            "type": "boolean",
            "title": "Auth Disabled"
          },
          "rules_version": {
            "type": "integer",
            "title": "Rules Version"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "service",
          "site_id",
          "machines",
          "mqtt",
          "sync",
          "models",
          "sim",
          "auth_disabled",
          "rules_version"
        ],
        "title": "SystemStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}